Added presenter mapper
