+++ src/Application/PresenterFactory.php
@@ -20,22 +20,20 @@ class PresenterFactory extends Nette\Object implements IPresenterFactory
 	/** @var bool */
 	public $caseSensitive = FALSE;
 
-	/** @var array[] of module => splited mask */
-	private $mapping = array(
-		'*' => array('', '*Module\\', '*Presenter'),
-		'Nette' => array('NetteModule\\', '*\\', '*Presenter'),
-	);
-
 	/** @var array */
 	private $cache = array();
 
 	/** @var Nette\DI\Container */
 	private $container;
 
+	/** @var \Nette\Application\IPresenterMapper */
+	private $presenterMapper;
 
-	public function __construct(Nette\DI\Container $container)
+
+	public function __construct(Nette\DI\Container $container, IPresenterMapper $presenterMapper)
 	{
 		$this->container = $container;
+		$this->presenterMapper = $presenterMapper;
 	}
 
 
@@ -78,7 +76,7 @@ class PresenterFactory extends Nette\Object implements IPresenterFactory
 			throw new InvalidPresenterException("Presenter name must be alphanumeric string, '$name' is invalid.");
 		}
 
-		$class = $this->formatPresenterClass($name);
+		$class = $this->presenterMapper->formatPresenterClass($name);
 		if (!class_exists($class)) {
 			throw new InvalidPresenterException("Cannot load presenter '$name', class '$class' was not found.");
 		}
@@ -93,7 +91,7 @@ class PresenterFactory extends Nette\Object implements IPresenterFactory
 		}
 
 		// canonicalize presenter name
-		$realName = $this->unformatPresenterClass($class);
+		$realName = $this->presenterMapper->unformatPresenterClass($class);
 		if ($name !== $realName) {
 			if ($this->caseSensitive) {
 				throw new InvalidPresenterException("Cannot load presenter '$name', case mismatch. Real name is '$realName'.");
@@ -106,58 +104,6 @@ class PresenterFactory extends Nette\Object implements IPresenterFactory
 		}
 
 		return $class;
-	}
-
-
-	/**
-	 * Sets mapping as pairs [module => mask]
-	 * @return self
-	 */
-	public function setMapping(array $mapping)
-	{
-		foreach ($mapping as $module => $mask) {
-			if (!preg_match('#^\\\\?([\w\\\\]*\\\\)?(\w*\*\w*?\\\\)?([\w\\\\]*\*\w*)\z#', $mask, $m)) {
-				throw new Nette\InvalidStateException("Invalid mapping mask '$mask'.");
-			}
-			$this->mapping[$module] = array($m[1], $m[2] ?: '*Module\\', $m[3]);
-		}
-		return $this;
-	}
-
-
-	/**
-	 * Formats presenter class name from its name.
-	 * @param  string
-	 * @return string
-	 */
-	public function formatPresenterClass($presenter)
-	{
-		$parts = explode(':', $presenter);
-		$mapping = isset($parts[1], $this->mapping[$parts[0]])
-			? $this->mapping[array_shift($parts)]
-			: $this->mapping['*'];
-
-		while ($part = array_shift($parts)) {
-			$mapping[0] .= str_replace('*', $part, $mapping[$parts ? 1 : 2]);
-		}
-		return $mapping[0];
-	}
-
-
-	/**
-	 * Formats presenter name from class name.
-	 * @param  string
-	 * @return string
-	 */
-	public function unformatPresenterClass($class)
-	{
-		foreach ($this->mapping as $module => $mapping) {
-			$mapping = str_replace(array('\\', '*'), array('\\\\', '(\w+)'), $mapping);
-			if (preg_match("#^\\\\?$mapping[0]((?:$mapping[1])*)$mapping[2]\\z#i", $class, $matches)) {
-				return ($module === '*' ? '' : $module . ':')
-					. preg_replace("#$mapping[1]#iA", '$1:', $matches[1]) . $matches[3];
-			}
-		}
 	}
 
 }
